+++ src/main/java/top/joylife/tracker/common/bean/param/OfferParam.java
@@ -2,6 +2,7 @@ package top.joylife.tracker.common.bean.param;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
@@ -30,7 +31,7 @@ public class OfferParam {
     /**
      * 支付金额
      */
-    private Integer payout;
+    private BigDecimal payout;
 
     /**
      * 备注
